pyproject.toml
[tool.poetry]
name = "blrec-auto-upload"
version = "0.1.0"
description = "自用版录播上传"
authors = ["qwerdvd <2450899274@qq.com>"]
readme = "README.md"
packages = [{include = "blrec_auto_upload"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.89.1"
uvicorn = "^0.20.0"
loguru = "^0.6.0"
python-dotenv = "^0.21.0"
aiohttp = "^3.8.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple"
default = true
